Extract getComponentKey function into an utils file
