+++ packages/core/src/utils/cms.ts
@@ -1,0 +1,10 @@
+import type { ComponentType } from 'react'
+
+export type ComponentTypeWithComponentKey<T> = ComponentType<T> & {
+  $componentKey?: string
+}
+
+export const getComponentKey = (
+  Component: ComponentTypeWithComponentKey<any>,
+  name: string
+) => Component.$componentKey ?? name
